@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:Wishy/components/default_button.dart';
+import 'package:Wishy/components/refound_terms.dart';
 import 'package:Wishy/constants.dart';
 import 'package:Wishy/size_config.dart';
 import 'package:flutter/gestures.dart';
@@ -7,14 +8,14 @@ import 'package:flutter/gestures.dart';
 class PaymentButton extends StatelessWidget {
   final onSubmit;
   final bool enable;
-  final bool loading;
   final double price;
   final Map<String, dynamic>? eventData;
+  final Widget? element;
 
   PaymentButton(
       {required this.price,
       required this.onSubmit,
-      required this.loading,
+      this.element,
       this.enable = true,
       this.eventData});
 
@@ -28,27 +29,36 @@ class PaymentButton extends StatelessWidget {
             style: Theme.of(context).textTheme.bodySmall,
             children: <TextSpan>[
               TextSpan(
-                text:
-                    "By completing this purchase, you confirm that you agree with our ",
+                text: "See our full ",
               ),
               TextSpan(
-                text: "Purchase and Refund Terms",
+                text: "Return Policy",
                 style: TextStyle(
-                  color: Colors.blue,
-                  decoration: TextDecoration.underline,
-                ),
-                recognizer: TapGestureRecognizer()..onTap = () {},
+                    color: Colors.blue,
+                    decoration: TextDecoration.underline,
+                    decorationColor: Colors.blue),
+                recognizer: TapGestureRecognizer()
+                  ..onTap = () {
+                    showDialog(
+                      context: context,
+                      barrierDismissible: false,
+                      builder: (BuildContext context) {
+                        return Dialog.fullscreen(child: RefoundTermsWebView());
+                      },
+                    );
+                  },
               ),
             ],
           ),
         ),
-        SizedBox(height: getProportionateScreenHeight(10)),
+        SizedBox(height: getProportionateScreenHeight(5)),
         DefaultButton(
-          text: "Confirm Purchase",
+          text: "Pay ${marketDetails["symbol"]}${price}",
           press: onSubmit,
           enable: enable,
           eventName: analyticEvents["PAY_PRESSED"],
           eventData: eventData,
+          element: element,
         ),
       ],
     );
