@@ -7,16 +7,20 @@ class TopRoundedContainer extends StatelessWidget {
     Key? key,
     required this.color,
     required this.child,
+    this.margin = 20,
+    this.padding = 20,
   }) : super(key: key);
 
   final Color color;
   final Widget child;
+  final double margin;
+  final double padding;
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      margin: EdgeInsets.only(top: getProportionateScreenWidth(20)),
-      padding: EdgeInsets.only(top: getProportionateScreenWidth(20)),
+      margin: EdgeInsets.only(top: getProportionateScreenWidth(margin)),
+      padding: EdgeInsets.only(top: getProportionateScreenWidth(padding)),
       width: double.infinity,
       decoration: BoxDecoration(
         color: color,
