@@ -1,17 +1,20 @@
+import 'package:Wishy/models/Product.dart';
 import 'package:flutter/material.dart';
 import '../../constants.dart';
 import '../../size_config.dart';
 
 class VariantPicker extends StatefulWidget {
   final String type;
-  final List<dynamic> values;
+  final List<Attribute> availableValues;
+  final List<Attribute> attributes;
   final String? chosenVariant;
-  final Function(String, String) onVariantChange;
+  final Function(Attribute) onVariantChange;
 
   const VariantPicker({
     Key? key,
     required this.type,
-    required this.values,
+    required this.availableValues,
+    required this.attributes,
     required this.onVariantChange,
     this.chosenVariant,
   }) : super(key: key);
@@ -26,8 +29,8 @@ class _VariantPickerState extends State<VariantPicker> {
   @override
   Widget build(BuildContext context) {
     if (widget.chosenVariant != null)
-      selectedIndex =
-          widget.values.indexWhere((item) => item == widget.chosenVariant);
+      selectedIndex = widget.attributes
+          .indexWhere((item) => item.value == widget.chosenVariant);
 
     return Padding(
       padding:
@@ -38,7 +41,7 @@ class _VariantPickerState extends State<VariantPicker> {
           Align(
               alignment: Alignment.centerLeft,
               child: Text(
-                "${widget.type[0].toUpperCase()}${widget.type.substring(1)}:",
+                widget.type,
               )),
           SizedBox(height: getProportionateScreenWidth(10)),
           Align(
@@ -47,21 +50,20 @@ class _VariantPickerState extends State<VariantPicker> {
                 spacing: 10,
                 runSpacing: 10,
                 children: List.generate(
-                  widget.values.length,
-                  (index) => GestureDetector(
-                    onTap: widget.chosenVariant == null
+                  widget.attributes.length,
+                  (index) => VariantElement(
+                    onSelect: widget.chosenVariant == null
                         ? () {
                             setState(() {
                               selectedIndex = index;
                             });
-                            widget.onVariantChange(
-                                widget.type, widget.values[index]);
+                            widget.onVariantChange(widget.attributes[index]);
                           }
                         : null,
-                    child: VariantDot(
-                      value: widget.values[index],
-                      isSelected: index == selectedIndex,
-                    ),
+                    enable: widget.attributes[index]
+                        .isExistIn(widget.availableValues),
+                    value: widget.attributes[index].value,
+                    isSelected: index == selectedIndex,
                   ),
                 ),
               )),
@@ -71,33 +73,56 @@ class _VariantPickerState extends State<VariantPicker> {
   }
 }
 
-class VariantDot extends StatelessWidget {
+class VariantElement extends StatelessWidget {
   final dynamic value;
   final bool isSelected;
+  final bool enable;
+  final void Function()? onSelect;
 
-  const VariantDot({
+  const VariantElement({
     Key? key,
     required this.value,
+    required this.enable,
+    required this.onSelect,
     this.isSelected = false,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.only(right: 10),
-      padding: EdgeInsets.symmetric(
-        horizontal: 16,
-        vertical: 6,
-      ),
-      decoration: BoxDecoration(
-        color: isSelected ? kPrimaryColor : Colors.transparent,
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: kPrimaryColor),
-      ),
-      child: Text(
-        value.toString(),
-        style: TextStyle(
-          color: isSelected ? Colors.white : kTextColor,
+    Color backgroundColor = isSelected
+        ? enable
+            ? kPrimaryColor
+            : kPrimaryColor.withOpacity(0.4)
+        : enable
+            ? Colors.transparent
+            : Colors.grey.withOpacity(0.5);
+
+    Color borderColor = enable ? kPrimaryColor : kSecondaryColor;
+
+    Color textColor = isSelected
+        ? Colors.white
+        : enable
+            ? kTextColor
+            : Colors.grey;
+
+    return GestureDetector(
+      onTap: onSelect,
+      child: Container(
+        margin: EdgeInsets.only(right: 10),
+        padding: EdgeInsets.symmetric(horizontal: 16, vertical: 6),
+        decoration: BoxDecoration(
+          color: backgroundColor,
+          borderRadius: BorderRadius.circular(8),
+          border: Border.all(color: borderColor),
+        ),
+        child: Opacity(
+          opacity: enable ? 1.0 : 0.5,
+          child: Text(
+            value.toString(),
+            style: TextStyle(
+              color: textColor,
+            ),
+          ),
         ),
       ),
     );
