@@ -1,16 +1,11 @@
-import 'package:Wishy/components/variants/variants_form.dart';
 import 'package:flutter/material.dart';
 import 'package:Wishy/components/variants/variants_widget.dart';
 import 'package:Wishy/models/Product.dart';
 import 'package:Wishy/size_config.dart';
 
-void showVariantsModal(BuildContext context, int productId, String productTitle,
-    List<Variant> variants, int? recipientId, String situation,
+void showVariantsModal(
+    BuildContext context, Product product, int? recipientId, String situation,
     {String? cursor}) {
-  final variantsObjects = groupVariants(variants);
-  int nonNullValueCount =
-      variantsObjects.values.where((value) => value != null).length;
-
   showModalBottomSheet<void>(
       context: context,
       isScrollControlled: true,
@@ -21,30 +16,26 @@ void showVariantsModal(BuildContext context, int productId, String productTitle,
         ),
       ),
       builder: (BuildContext context) {
-        return FractionallySizedBox(
-            heightFactor: nonNullValueCount > 2 ? 0.72 : 0.5,
+        return SingleChildScrollView(
             child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              mainAxisSize: MainAxisSize.max,
-              children: <Widget>[
-                Container(
-                    padding:
-                        EdgeInsets.only(top: getProportionateScreenWidth(20)),
-                    child: Text(
-                      'Select Your Preferred Variant',
-                      style: TextStyle(
-                        fontSize: getProportionateScreenWidth(18),
-                      ),
-                    )),
-                VariantsWidget(
-                  productId: productId,
-                  productTitle: productTitle,
-                  productVariants: variants,
-                  recipientId: recipientId,
-                  situation: situation,
-                  cursor: cursor,
-                )
-              ],
-            ));
+          mainAxisAlignment: MainAxisAlignment.center,
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            Container(
+                padding: EdgeInsets.only(top: getProportionateScreenWidth(20)),
+                child: Text(
+                  'Select Your Preferred Variant',
+                  style: TextStyle(
+                    fontSize: getProportionateScreenWidth(18),
+                  ),
+                )),
+            VariantsWidget(
+              product: product,
+              recipientId: recipientId,
+              situation: situation,
+              cursor: cursor,
+            )
+          ],
+        ));
       });
 }
