@@ -1,8 +1,4 @@
 import 'dart:async';
-
-import 'package:Wishy/services/opt_services.dart';
-import 'package:Wishy/utils/notification.dart';
-import 'package:Wishy/utils/router_utils.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
@@ -14,12 +10,13 @@ import 'package:Wishy/screens/root_screen.dart';
 import 'package:Wishy/theme.dart';
 import 'package:Wishy/global_manager.dart';
 import 'package:Wishy/utils/analytics.dart';
-import 'package:uni_links/uni_links.dart';
+import 'package:flutter_stripe/flutter_stripe.dart';
 import 'package:uuid/uuid.dart';
 import 'firebase_options.dart';
+import 'package:provider/provider.dart';
 
 Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
-  // If you're going to use other Firebase services in the background, such as Firestore,
+  // If you're going to use other Firebase services in the background,
   // make sure you call `initializeApp` before using other Firebase services.
   print("Handling a background message: ${message.messageId}");
 }
@@ -43,63 +40,16 @@ void main() async {
   await GlobalManager().initialize();
   final uuid = Uuid();
   GlobalManager().setSession(uuid.v1());
-  final authServices = AuthServices();
+  Stripe.publishableKey = dotenv.get("STRIPE_KEY");
 
   await AnalyticsService.init({"User Id": GlobalManager().userId});
-  updateNotificationPermission();
 
-  try {
-    final userLocation = await authServices.userLocationData();
-    GlobalManager().setUserLocation(userLocation);
-
-    if (userLocation.isProductsAvailable) {
-      GlobalManager().setDeliveryAvailability(true);
-    }
-  } catch (error) {
-    print(error);
-    GlobalManager().setDeliveryAvailability(true);
-  }
-
-  runApp(MyApp());
-  initUniLinks();
-}
-
-Future<void> initUniLinks() async {
-  try {
-    StreamSubscription? _linkSubscription;
-
-    _linkSubscription = uriLinkStream.listen((Uri? uri) {
-      handleDeepLink(uri);
-    }, onError: (err) {
-      print('Error on uriLinkStream: $err');
-    });
-
-    await getInitialUri().then((value) => handleDeepLink(value));
-  } catch (e) {
-    print('Error initializing deep links: $e');
-  }
-}
-
-void handleDeepLink(Uri? uri) {
-  if (uri != null) {
-    String? navigationToken = null;
-
-    print('Handling URI: $uri');
-    if (uri.path == '/requests') {
-      GlobalManager().setNavigateToRequest(true);
-      navigationToken = uri.queryParameters["token"];
-    } else if (uri.path == '/invites') {
-      navigationToken = uri.queryParameters["token"];
-    } else {
-      print("Path not handled: ${uri.path}");
-    }
-
-    GlobalManager().setNotificationToken(navigationToken);
-  }
-
-  GlobalManager().setShowAnimation(GlobalManager().token == null);
-
-  RouterUtils.routeToHomePage();
+  runApp(
+    ChangeNotifierProvider(
+      create: (context) => GlobalManager(),
+      child: MyApp(),
+    ),
+  );
 }
 
 class MyApp extends StatelessWidget {
