@@ -27,7 +27,7 @@ final Map<String, WidgetBuilder> routes = {
     final args = ModalRoute.of(context)!.settings.arguments as Map;
     return CheckoutScreen(
         variant: args['variant'] as Variant,
-        productId: args['productId'] as int,
+        product: args['product'] as Product,
         recipientId: args['recipientId'] as int?);
   },
   HomeScreen.routeName: (context) => HomeScreen(),
