@@ -1,34 +1,32 @@
+import 'package:Wishy/screens/checkout/components/product_widget.dart';
 import 'package:Wishy/screens/checkout/components/purchase_form.dart';
 import 'package:flutter/material.dart';
-import 'package:Wishy/constants.dart';
 import 'package:Wishy/models/Product.dart';
-import 'package:Wishy/size_config.dart';
 
 class Body extends StatelessWidget {
   final Variant variant;
-  final int productId;
+  final Product product;
   final int? recipientId;
 
-  Body({required this.variant, required this.productId, this.recipientId});
+  Body({required this.variant, required this.product, this.recipientId});
 
   @override
   Widget build(BuildContext context) {
     return SafeArea(
         child: SingleChildScrollView(
-            child: Padding(
-      padding:
-          EdgeInsets.symmetric(horizontal: getProportionateScreenWidth(20)),
       child: Column(
         children: [
-          Text("Complete Purchase", style: headingStyle),
-          SizedBox(height: getProportionateScreenHeight(10)),
+          ProductWidget(
+            product: product,
+            variant: variant,
+          ),
           PurchaseForm(
-              productId: productId,
+              product: product,
               variantId: variant.id,
               recipientId: recipientId,
               price: variant.price),
         ],
       ),
-    )));
+    ));
   }
 }
