@@ -3,6 +3,7 @@ import 'package:Wishy/global_manager.dart';
 import 'package:Wishy/screens/checkout/checkout_screen.dart';
 import 'package:Wishy/utils/analytics.dart';
 import 'package:Wishy/utils/is_variants_exists.dart';
+import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
 import 'package:Wishy/components/additional_details_dialog.dart';
 import 'package:Wishy/components/default_button.dart';
@@ -15,40 +16,104 @@ import 'product_description.dart';
 import '../../../components/top_rounded_container.dart';
 import 'product_images.dart';
 
-class Body extends StatelessWidget {
-  final situation = "product_details";
+class Body extends StatefulWidget {
   final Product product;
   final int? variantId, recipientId;
   final String buttonText;
   final String? cursor;
 
-  Body(
-      {Key? key,
-      required this.product,
-      this.buttonText = "Buy Now",
-      this.variantId,
-      this.cursor,
-      this.recipientId})
-      : super(key: key);
+  Body({
+    Key? key,
+    required this.product,
+    this.buttonText = "Buy Now",
+    this.variantId,
+    this.cursor,
+    this.recipientId,
+  }) : super(key: key);
+
+  @override
+  _BodyState createState() => _BodyState();
+}
+
+class _BodyState extends State<Body> {
+  final situation = "product_details";
+  int? _selectedImageIndex = null;
+  ScrollController _scrollController = ScrollController();
+  CarouselController _carouselController = CarouselController();
+
   final firstColor = Colors.white;
   final secondColor = Color(0xFFF6F7F9);
+
+  void _sendImageViewedEvent() {
+    AnalyticsService.trackEvent(analyticEvents["PRODUCT_IMAGE_VIEWED"]!,
+        properties: {
+          'Product Id': widget.product.id,
+          'Product Title': widget.product.title,
+          'Image Url': widget.product.images.isNotEmpty
+              ? widget.product.images[_selectedImageIndex ?? 0].url
+              : null,
+          "Situation": situation
+        });
+  }
+
+  void _onImageChange(int index) {
+    setState(() {
+      _selectedImageIndex = index;
+    });
+
+    double position = index * (getProportionateScreenWidth(48) + 15);
+    _scrollController.animateTo(
+      position,
+      duration: defaultDuration,
+      curve: Curves.easeInOut,
+    );
+    _sendImageViewedEvent();
+  }
+
+  void _onVariantChange(Variant? v) {
+    int index = 0;
+    if (v == null) return;
+
+    if (v.imageId != null) {
+      index = widget.product.images
+          .indexWhere((element) => element.id == v.imageId);
+    }
+
+    _carouselController.jumpToPage(index);
+    _onImageChange(index);
+  }
+
+  @override
+  void initState() {
+    if (_selectedImageIndex == null)
+      Future.delayed(
+          Duration.zero, () => _onVariantChange(widget.product.variants![0]));
+
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
     return ListView(
       children: [
-        ProductImages(product: product),
+        ProductImages(
+          carouselController: _carouselController,
+          scrollController: _scrollController,
+          onImageChange: _onImageChange,
+          selectedImage: _selectedImageIndex ?? 0,
+          product: widget.product,
+        ),
         TopRoundedContainer(
           color: Colors.white,
           child: Column(
             children: [
               ProductDescription(
-                product: product,
+                product: widget.product,
                 pressOnSeeMore: () {
                   AnalyticsService.trackEvent(
                       analyticEvents["SHOW_MORE_PRODUCT_DESCRIPTION"]!,
                       properties: {
-                        "Product Id": product.id,
+                        "Product Id": widget.product.id,
                         "Delivery Availability":
                             GlobalManager().isDeliveryAvailable
                       });
@@ -57,23 +122,22 @@ class Body extends StatelessWidget {
                     context: context,
                     builder: (BuildContext context) {
                       return AdditionalDetailsDialog(
-                        description: product.description ?? "",
+                        description: widget.product.description ?? "",
                       );
                     },
                   );
                 },
               ),
-              if (product.variants == null)
+              if (widget.product.variants == null)
                 buildOutOfStock(secondColor)
-              else if (isVariantsExists(product.variants))
+              else if (isVariantsExists(widget.product.variants))
                 VariantsWidget(
+                    onVariantChange: _onVariantChange,
                     situation: situation,
-                    productId: product.id,
-                    productTitle: product.title,
-                    productVariants: product.variants!,
-                    buttonText: this.buttonText,
-                    variantId: this.variantId,
-                    recipientId: recipientId)
+                    product: widget.product,
+                    buttonText: widget.buttonText,
+                    variantId: widget.variantId,
+                    recipientId: widget.recipientId)
               else
                 buildButton(secondColor, context)
             ],
@@ -95,11 +159,11 @@ class Body extends StatelessWidget {
         ),
         child: DefaultButton(
           text:
-              "${this.buttonText} ${marketDetails["symbol"]}${product.variants![0].price}",
+              "${widget.buttonText} ${marketDetails["symbol"]}${widget.product.variants![0].price}",
           eventName: analyticEvents["CHECKOUT_PRESSED"]!,
           eventData: {
-            "Product Id": product.id,
-            "Product Title": product.title,
+            "Product Id": widget.product.id,
+            "Product Title": widget.product.title,
             "Situation": situation,
             "Variants Exist": false,
             "Variant Picked": false,
@@ -116,10 +180,10 @@ class Body extends StatelessWidget {
               context,
               CheckoutScreen.routeName,
               arguments: {
-                'variant': product.variants![0],
-                'productId': product.id,
-                'recipientId': recipientId,
-                "cursor": cursor
+                'variant': widget.product.variants![0],
+                'productId': widget.product.id,
+                'recipientId': widget.recipientId,
+                "cursor": widget.cursor
               },
             );
           },
