@@ -31,7 +31,7 @@ class ProfileMenu extends StatelessWidget {
           children: [
             SvgPicture.asset(
               icon,
-              color: kPrimaryColor,
+              colorFilter: ColorFilter.mode(kPrimaryColor, BlendMode.srcIn),
               width: 22,
             ),
             SizedBox(width: 20),
