@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_initicon/flutter_initicon.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:Wishy/global_manager.dart';
 
 class ProfilePic extends StatelessWidget {
@@ -14,7 +13,7 @@ class ProfilePic extends StatelessWidget {
       height: 80,
       width: 80,
       child: Initicon(
-        text: GlobalManager().username ?? "",
+        text: GlobalManager().user?.name ?? "",
         elevation: 4,
       ),
     );
