@@ -5,12 +5,7 @@ bool isVariantsExists(List<Variant>? variants) {
     return false;
   else if (variants.length == 1) {
     final variant = variants[0];
-    if ((variant.color == null || variant.color == "") &&
-        (variant.size == null || variant.size == "") &&
-        (variant.material == null || variant.material == "") &&
-        (variant.style == null || variant.style == "")) {
-      return false;
-    }
+    return variant.attributes != null;
   }
 
   return true;
